Create Tower_MeleeBurningMadness
A tower that release an area of effect that causes low burning damage and causes frenzy to the targets that are affected. The area was reduced to 10ft of the tower, due to an experience where half of the enemy spawn horde fought each other to the d**th while the player stood and watched. 

each rank increases the duration of the fire and frenzy by 1 second. Rank 5 lasts indefinitely