Create Tower_MeleeGreenmote
A large pile of greenmote, a powerful hallucinogen consumed by Thadon, which ultimately leads to his d**th in the game. Every 10 seconds, it's vapors paralyzes a target for 1 second and causes them to suffer 5 cold damage until it is dispelled. This is to mimic the effects of withdrawal that Thadon suffers from frequently in the game. 

each Rank adds +1 cold damage. Rank 5 drains the targets' speed by 20 points and paralyzes for 2 seconds.