Create Unit_MeleeTreasureGoblin
A golden goblin that maintains proximity to the player and attacks enemies. Each time they hit a target, the player gains +1 gold (game currency). Each rank adds +1 gold per hit. At Rank 5, the goblins cannot be killed.

there is a 0.01% chance on hit that the goblin gives the player an unidentified item on hit.