Create Hero_OrcCaptainAdventurer
At the beginning of the game, an orc captain rushes the gates of madness with 5 adventurers, all clad in steel armor. They are killed by Gatekeeper Thanos and their bodies littter the walkway to the gates.

This hero can be summoned with his whole squad. Although they are not very powerful, they come in numbers and can protect the player. They cannot be killed, but can be knocked unconscious.