Create Hero_ArdenSulHaskill
Throughout the story mode of Sheogorath / The Shivering Isles, the player can find multiple relics belonging to a former champion of Sheogorath called Arden-Sul, but nothing in the wiki or known reddit-posts theorize who he is. 

SO I set some pieces in place that reveal that to be Haskil. Essentially, you find his sword and when you return to Haskil, he will ask you why you bring it to him, then reveals that the blade belongs to him and was left at the site of his defeat to Jyggalag in the previous era. With this blade returned, he will join you in defense of the realm.

Arden-Sul / Haskil is a powerful offense with -% defense. Basically, the less defense he has (weakness to elements), the more damage he deals (each -1% adds +1 damage to his weapon). His attacks heal him 2% of his total health and he regains full health on kill.